database/requirements.txt
my-sql-connector
kagglehub
pandas
np
beautifulsoup4
